Fix fermata pasting, update tests

+++ src/engraving/editing/cmd.cpp
@@ -1262,7 +1262,7 @@ Segment* Score::setNoteRest(Segment* segment, track_idx_t track, NoteVal nval, F
 //    return size of actual gap
 //---------------------------------------------------------
 
-Fraction Score::makeGap(Segment* segment, track_idx_t track, const Fraction& _sd, Tuplet* tuplet, bool keepChord)
+Fraction Score::makeGap(Segment* segment, track_idx_t track, const Fraction& _sd, Tuplet* tuplet, bool keepChord, bool deleteAnnotations)
 {
     assert(_sd.numerator());
 
@@ -1432,14 +1432,16 @@ Fraction Score::makeGap(Segment* segment, track_idx_t track, const Fraction& _sd
         filter.setFiltered(ElementsSelectionFilterTypes::CHORD_SYMBOL, false);
         filter.setFiltered(ElementsSelectionFilterTypes::FRET_DIAGRAM, false);
 
-        deleteAnnotationsFromRange(s1, s2, track, track + 1, filter);
+        if (deleteAnnotations) {
+            deleteAnnotationsFromRange(s1, s2, track, track + 1, filter);
+        }
         deleteSlursFromRange(t1, t2, track, track + 1, filter);
     }
 
     return accumulated;
 }
 
-bool Score::makeGapVoice(Segment* seg, track_idx_t track, Fraction len, const Fraction& tick)
+bool Score::makeGapVoice(Segment* seg, track_idx_t track, Fraction len, const Fraction& tick, bool deleteAnnotations)
 {
     ChordRest* cr = 0;
     cr = toChordRest(seg->element(track));
@@ -1453,7 +1455,7 @@ bool Score::makeGapVoice(Segment* seg, track_idx_t track, Fraction len, const Fr
                 }
                 // this happens only for voices other than voice 1
                 expandVoice(seg, track);
-                return makeGapVoice(seg, track, len, tick);
+                return makeGapVoice(seg, track, len, tick, deleteAnnotations);
             }
             if (seg1->element(track)) {
                 break;
@@ -1512,7 +1514,7 @@ bool Score::makeGapVoice(Segment* seg, track_idx_t track, Fraction len, const Fr
             LOGD("cannot make gap");
             return false;
         }
-        Fraction l = makeGap(cr->segment(), cr->track(), len, 0);
+        Fraction l = makeGap(cr->segment(), cr->track(), len, nullptr, /*keepChord*/ false, deleteAnnotations);
         if (l.isZero()) {
             LOGD("returns zero gap");
             return false;
